Remove ssl_prefer_server_ciphers nginx option
This option was useful with TLS versions older than 1.2 to ensure that
browsers would not pick outdated ciphers without forward secrecy.

With the current cipher list we have configured on TLS 1.2 and newer
this is no longer a problem. Allowing the client to pick cipher is a
performance optimization, since it means that phones without AES
accelleration can choose to use the faster CHACHA20-POLY1305 mode.