Fix haproxy config, use ALPN for content switching
HTTP2 is a binary protocol. So switching on the first byte does not
work. If the client wants to speak HTTP2, it will negotiate this by ALPN
in the TLS handshake. haproxy can use this to send traffic in the right
direction.

As of section 3.1 in RFC 7540 using APLN to negotiate on HTTP2 is
mandatory for encrypted connections.