fix permissions, add UDP TURN config
TURN connections should be offered on two transports to the client:

* port 443 TCP (thats what haproxy redirects to coturn)
* port 3478 UDP

The UDP port likely offers better quality for clients than the TCP
connections as TCP connections stall when packets are lost. So UDP
connections should be always preferred over TCP connections for WebRTC
connections as long as udp is not blocked by firewalls. Firefox needs to
relay all its traffic that goes to mediasoup through a turn server as a
workaround for their broken ICE implementation.

With fullaudio which lets mediasoup proxy audio connections to
freeswitch this is becoming even more important for a good out of the
box experience.