Fix TURN server configuration for TURN servers behind NAT gateways
Commit 1b19b0 (Updated for installing coturn on 20.04 - Jan 17 2021) added
coturn configuration options (listening-ip, relay-ip, external-ip) that
aren't properly set if the TURN server is behind a NAT gateway.

In the NAT case, both listening-ip (the address the TURN server binds to)
and relay-ip (the address the TURN server transmits on) should be the
private IP address.

If listening-ip and relay-ip are improperly set to the external IP address,
then we get error messages (like those reported on Issue #577) that the
TURN server can't bind to the public IP address.

Fixes and closes Issue #577.