Set $real_scheme variable in nginx config
Updates to BigBlueButton in
https://github.com/bigbluebutton/bigbluebutton/pull/17732
require that the /etc/nginx/sites-available/bigbluebutton file sets the
$real_scheme variable as appropriate for the listener and reverse proxy
situation.

When nginx is terminating ssl itself, set it to the value of $scheme,
same as the config file shipped with BigBlueButton.
When nginx is being reverse-proxied behind haproxy, set this variable to
the fixed string "https" since nginx does not have any way of knowing
that the original connection had been made using TLS.

A couple of stray formatting changes (newlines, trailing spaces) snuck
in to sync the bbb 2.6 and 2.7 versions of the script to each-other.