Fix issue #789
There is a conflict between UFW and iptables-persistent / netfilter-persistent, due to a known bug in Ubuntu 22.04 https://bugs.launchpad.net/ufw/+bug/1987227 After reboot with UFW enabled, the network connectivity breaks because every single package is dropped. I was able to circumvent this bug by modifying BBB installation script so that iptables-persistent is only being installed if UFW is NOT being installed. If UFW is being installed, I'm trying to mimic iptables-persistent functionality with UFW.