Update ImageMagick policy to allow the XC module
The presentation format recording scripts use the xc module (which
stands for "Constant Color" to generate placeholder images for slides
which are missing / failed to convert and a few other circumstances.

The XC module is safe to enable, since it doesn't involve reading any
files - it only generates single-color in memory images, and image sizes
(memory usage) are limited by other sections of the policy.